requirements.txt
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.6
django-heroku==0.3.1
djangorestframework==3.10.3
gunicorn==19.9.0
idna==2.8
kiwisolver==1.1.0
matplotlib==3.1.1
nltk==3.4.5
numpy==1.17.2
oauthlib==3.1.0
pandas==0.25.1
pickle-mixin==1.0.2
Pillow==6.2.0
plotly==4.2.1
psycopg2==2.8.3
PyJWT==1.7.1
pyparsing==2.4.2
PySocks==1.7.0
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
retrying==1.3.3
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
textblob==0.15.3
tweepy==3.8.0
urllib3==1.25.3
whitenoise==4.1.3
